# contracts/openapi/v1/openapi.yaml
openapi: 3.0.3
info:
  title: GoldPC API
  description: |
    REST API для веб-приложения компьютерного магазина GoldPC с сервисным центром.
    
    ## Аутентификация
    API использует JWT-токены для аутентификации. Включите токен в заголовок:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Версионирование
    Текущая версия API: v1. Версионирование через URL путь (/api/v1/...).
    
    ## Формат дат
    Все даты и время в формате ISO 8601 (UTC).
    
    ## Коды ошибок
    - `400` - Некорректный запрос
    - `401` - Неавторизован
    - `403` - Доступ запрещён
    - `404` - Ресурс не найден
    - `409` - Конфликт
    - `422` - Ошибка валидации
    - `429` - Превышен лимит запросов
    - `500` - Внутренняя ошибка сервера
    
    ## Роли пользователей
    - **Client** — Клиент (оформление заказов, просмотр каталога)
    - **Manager** — Менеджер (обработка заказов, управление товарами)
    - **Master** — Мастер (выполнение услуг сервисного центра)
    - **Admin** — Администратор (полный доступ)
    - **Accountant** — Бухгалтер (финансовые отчёты)
  version: 1.0.0
  contact:
    name: GoldPC API Support
    email: api@goldpc.by
    url: https://goldpc.by/api-docs
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.goldpc.by/v1
    description: Production server
  - url: https://api-staging.goldpc.by/v1
    description: Staging server
  - url: http://localhost:5000/v1
    description: Development server
...
security:
  - bearerAuth: []
paths:
...
  /orders:
    get:
      tags: [Orders]
      summary: Получить список заказов
      operationId: getOrders
      parameters:
        - $ref: './components/parameters/common.yaml#/components/parameters/Page'
        - $ref: './components/parameters/common.yaml#/components/parameters/PageSize'
        - $ref: './components/parameters/common.yaml#/components/parameters/Status'
        - $ref: './components/parameters/common.yaml#/components/parameters/DateFrom'
        - $ref: './components/parameters/common.yaml#/components/parameters/DateTo'
        - $ref: './components/parameters/common.yaml#/components/parameters/Search'
      responses:
        '200':
          description: Список заказов
          content:
            application/json:
              schema:
                $ref: './components/schemas/order.yaml#/components/schemas/OrderListResponse'

    post:
      tags: [Orders]
      summary: Создать заказ
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas/order.yaml#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Заказ создан
          content:
            application/json:
              schema:
                $ref: './components/schemas/order.yaml#/components/schemas/Order'
        '400':
          $ref: './components/responses/common.yaml#/components/responses/ValidationError'
...
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Получить заказ по ID
      operationId: getOrder
      parameters:
        - $ref: './components/parameters/common.yaml#/components/parameters/OrderId'
      responses:
        '200':
          description: Детали заказа
          content:
            application/json:
              schema:
                $ref: './components/schemas/order.yaml#/components/schemas/Order'
        '404':
          $ref: './components/responses/common.yaml#/components/responses/NotFound'
  /orders/{orderId}/cancel:
    post:
      tags: [Orders]
      summary: Отменить заказ
      operationId: cancelOrder
      parameters:
        - $ref: './components/parameters/common.yaml#/components/parameters/OrderId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: './components/schemas/order.yaml#/components/schemas/CancelOrderRequest'
      responses:
        '200':
          description: Заказ отменён
          content:
            application/json:
              schema:
                $ref: './components/schemas/order.yaml#/components/schemas/Order'
        '400':
          $ref: './components/responses/common.yaml#/components/responses/BadRequest'
        '404':
          $ref: './components/responses/common.yaml#/components/responses/NotFound'
  /orders/{orderId}/status:
    put:
      tags: [Orders]
      summary: Обновить статус заказа (Manager+)
      operationId: updateOrderStatus
      parameters:
        - $ref: './components/parameters/common.yaml#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas/order.yaml#/components/schemas/UpdateOrderStatusRequest'
      responses:
        '200':
          description: Статус обновлён
          content:
            application/json:
              schema:
                $ref: './components/schemas/order.yaml#/components/schemas/Order'
        '400':
          $ref: './components/responses/common.yaml#/components/responses/BadRequest'
        '403':
          $ref: './components/responses/common.yaml#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен авторизации